+++ Plugins/ags_snowrain/ags_snowrain.cpp
@@ -812,7 +812,7 @@ void AGS_EngineShutdown()
   delete snow;
 }
 
-int AGS_EngineOnEvent(int event, int data)
+int AGS_EngineOnEvent(int event, long data)
 {
   if (event == AGSE_PREGUIDRAW)
   {
